requirements.txt
qiskit==0.46.0
qiskit-aer==0.13.0
tensorflow==2.15.0
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
krakenex==2.1.0
requests==2.31.0
asyncio==3.4.3
transformers==4.38.2
fastapi==0.115.0
uvicorn==0.30.6
cryptography==43.0.1